requirements.txt
Flask==2.2.2
mysqlclient
flask_mysqldb
python-Levenshtein